refactor: reorganize pages structure
