+++ src/pages/Home/Home.tsx
@@ -1,17 +1,7 @@
-import { Carousel, Col, Divider, Input, Radio, Row } from "antd";
-import { slides } from "../assets/images/slides";
+import { Button, Carousel, Col, Divider, Input, Radio, Row } from "antd";
+import { slides } from "../../assets/images/slides";
 import type { CheckboxGroupProps } from "antd/es/checkbox";
-
-const contentStyle: React.CSSProperties = {
-  margin: 0,
-  textAlign: "center",
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  justifyContent: "center",
-  position: "relative",
-  backgroundColor: "#fff",
-};
+import "./Home.css";
 
 const options: CheckboxGroupProps<string>["options"] = [
   { label: "Name", value: "Name" },
@@ -19,28 +9,28 @@ const options: CheckboxGroupProps<string>["options"] = [
   { label: "Species", value: "Species" },
 ];
 
+const handleSearch = () => {
+  console.log("search");
+};
+
 export function Home() {
   return (
     <>
       <Row
         justify="center"
         align="middle"
-        style={{ minHeight: "90vh" }}
+        className="home-row"
         gutter={{ xs: 8, sm: 16, md: 24, lg: 32 }}
       >
         <Col span={6}>
           <Carousel autoplay dots={false} effect="fade" autoplaySpeed={4000}>
             {slides.map((slide, index) => (
               <div key={index}>
-                <div style={contentStyle}>
+                <div className="carousel-content">
                   <img
                     src={slide.image}
                     alt={slide.name}
-                    style={{
-                      width: "100%",
-                      objectFit: "cover",
-                      height: "500px",
-                    }}
+                    className="carousel-image"
                   />
                   <div>
                     <h2>{slide.name}</h2>
@@ -51,15 +41,21 @@ export function Home() {
           </Carousel>
         </Col>
         <Col span={6}>
-          <h3>Search by:</h3>
-          <Input size="large" />
-          <Divider variant="dashed" />
-          <Radio.Group
-            block
-            options={options}
-            defaultValue="Name"
-            optionType="button"
-          />
+          <div className="search-container">
+            <h3>Search by:</h3>
+            <Input size="large" />
+            <Divider variant="dashed" />
+            <Radio.Group
+              block
+              options={options}
+              defaultValue="Name"
+              optionType="button"
+            />
+
+            <Button type="default" size="large" onClick={handleSearch}>
+              Search
+            </Button>
+          </div>
         </Col>
       </Row>
     </>
